Send debugging information over the serial port rather than HID - TESTING ONLY
Send debugging information over the serial port rather than HID.  This
allows commands to be sent faster using bulk endpoints rather than
interrupt endpoints, which are limited to 64KB/s.  This patch also
bumps the CDC buffers so multiple packets can be received/sent.
This patch is not meant to be merged as is.